ProyectoIAMelanoma-FrontEnd/WEB/requirements.txt
fastapi
uvicorn
Pillow
torch
torchvision
numpy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
text

Añado test